requirements.txt
streamlit
pandas
numpy 
matplotlib
seaborn
scikit-learn
shap
xgboost
plotly
Pillow
streamlit-shap
